-Adicionado validação para criação dos relatórios e corrigidos alguns problemas com o relatório de andamento.

+++ src/main/java/com/fluxo/report/dto/FinalReportResponseDto.java
@@ -1,11 +1,12 @@
 package com.fluxo.report.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record FinalReportResponseDto(
         LocalDate date,
         String project,
-        double grade,
+        BigDecimal grade,
         String feedback
 ) {
 }
