+++ src/main/java/com/fluxo/report/dto/ProgressReportResponseDto.java
@@ -1,11 +1,12 @@
 package com.fluxo.report.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ProgressReportResponseDto(
         LocalDate date,
         String project,
-        double grade,
+        BigDecimal grade,
         String feedback
 ) {
 }
